requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
asyncpg==0.29.0
tortoise-orm==0.20.0
aerich==0.7.2